backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.1
gunicorn>=20.1.0
google-cloud-vision
google-cloud-texttospeech
google-cloud-speech
google-genai
python-dotenv==1.0.0
kokoro>=0.9.2
soundfile
torch==2.8.0
docker
pyphen
pdf2image
pillow